adjust to pull request comment
